+++ src/main/java/com/member/dao/MemberDAO.java
@@ -68,16 +68,6 @@ public interface MemberDAO {
 	public int countPosts(String userId) throws Exception;
 	public int countComments(String userId) throws Exception;
 
-	// admin
-	public List<MemberVO> listAllMembers(Criteria criteria);
-	public List<CollectionReviewVO> listAllReviews(Criteria criteria);
-	public List<BoardVO> listAllPosts(Criteria criteria);
-	public List<ReplyVO> listAllComments(Criteria criteria);
 
-	// admin PageMaker
-	public int countAllMembers() throws Exception;
-	public int countAllCollectionRevs() throws Exception;
-	public int countAllPosts() throws Exception;
-	public int countAllComments() throws Exception;
 
 }
