+++ src/main/java/com/member/dao/MemberDAOImpl.java
@@ -158,54 +158,6 @@ public class MemberDAOImpl implements MemberDAO {
         return sql.selectOne("memberMapper.countComments", userId);
     }
     
-    // admin
-    @Override
-    public List<MemberVO> listAllMembers(Criteria criteria) {
-    	Map<String, Object> paramMap = new HashMap<>();
-    	paramMap.put("start", criteria.getStart());
-      return sql.selectList("memberMapper.listAllMembers", paramMap);
-    }
-
-    @Override
-    public List<CollectionReviewVO> listAllReviews(Criteria criteria) {
-    	Map<String, Object> paramMap = new HashMap<>();
-    	paramMap.put("start", criteria.getStart());
-      return sql.selectList("memberMapper.listAllReviews", paramMap);
-    }
-
-    @Override
-    public List<BoardVO> listAllPosts(Criteria criteria) {
-    	Map<String, Object> paramMap = new HashMap<>();
-    	paramMap.put("start", criteria.getStart());
-      return sql.selectList("memberMapper.listAllPosts", paramMap);
-    }
-
-    @Override
-    public List<ReplyVO> listAllComments(Criteria criteria) {
-    	Map<String, Object> paramMap = new HashMap<>();
-    	paramMap.put("start", criteria.getStart());
-      return sql.selectList("memberMapper.listAllComments", paramMap);
-    }
-    
-    // admin PageMaker
-    @Override
-    public int countAllMembers() throws Exception {
-        return sql.selectOne("memberMapper.countAllMembers");
-    }
-    
-    @Override
-    public int countAllCollectionRevs() throws Exception {
-        return sql.selectOne("memberMapper.countAllCollectionRevs");
-    }
-
-    @Override
-    public int countAllPosts() throws Exception {
-        return sql.selectOne("memberMapper.countAllPosts");
-    }
-
-    @Override
-    public int countAllComments() throws Exception {
-        return sql.selectOne("memberMapper.countAllComments");
-    }
+   
 
 }
