+++ src/main/java/com/member/service/MemberService.java
@@ -54,16 +54,8 @@ public interface MemberService {
 	public int countPosts(String userId) throws Exception;
 	public int countComments(String userId) throws Exception;
 
-	// admin
-	public List<MemberVO> listAllMembers(Criteria criteria);
-	public List<CollectionReviewVO> listAllReviews(Criteria criteria);
-	public List<BoardVO> listAllPosts(Criteria criteria);
-	public List<ReplyVO> listAllComments(Criteria criteria);
 	
-	// admin PageMaker
-	public int countAllMembers() throws Exception;
-	public int countAllCollectionRevs() throws Exception;
-	public int countAllPosts() throws Exception;
-	public int countAllComments() throws Exception;
+
+	
 
 }
